Add support for "extra conf" startup options
We need this to be able to stop bin logs and other
configuration settings, because MySQL 8.0 comes with
it enabled by defaults and that eats the tmpfs
really quickly for big test suites.

This is built on top of https://github.com/mirromutth/mysql-action/pull/6
that has been added to our fork already.